requirements.txt
mistune
numpy
python-dotenv
requests
tqdm
tiktoken
waybackpy
rich
